Soft scroll-fade hints on Settings and onboarding too
The previous PR introduced the top/bottom gradient fades on the Today
pager pages — a soft hint that there's more to scroll. Apply the same
affordance everywhere else the user lands on a scrollable list: the
Settings root list, each Settings sub-page, and the onboarding flow.

Lift `EdgeFadeOverlay` out of `TodayScreen.kt` into a shared composable
at `ui/EdgeFadeOverlay.kt` so it isn't a Today-specific helper any
more. Add a `ScrollState` overload that wires `canScrollBackward` /
`canScrollForward` itself, cutting the call site to a single line in
each screen.

Each affected `*Content` composable hoists its `rememberScrollState()`,
wraps its scrolling `Column` with `EdgeFadeOverlay(scrollState = …)`,
and applies the Scaffold padding on the overlay so the fades sit at
the visible viewport edges (just below the TopAppBar at the top, at
the bottom of the screen below).

Affected screens: Settings root list, Schedule, Clothes, Region, Voice,
Display, Location, Forecasters, Calendar, SmartHome, Privacy, About,
plus the onboarding flow.

Snapshot coverage: Roborazzi auto-regenerated 10 PNGs under
`app/snapshots/` — the ones with enough content to overflow (Settings
root, Schedule, Clothes, Voice, Location, plus the three onboarding
states). Pages that fit on a single screen (Region, Display, Calendar,
Privacy, Forecasters, SmartHome) show no fade in their snapshots,
which is exactly the intended behaviour.
